Refactoring in an attempt to increase coverage.
