+++ gpflux/architectures/factory.py
@@ -35,7 +35,8 @@ from gpflux.helpers import (
     construct_basic_kernel,
     construct_mean_function,
 )
-from gpflux.layers.gp_layer import GPLayer, OrthGPLayer
+from gpflux.layers import OrthGPLayer
+from gpflux.layers.gp_layer import GPLayer
 from gpflux.layers.likelihood_layer import LikelihoodLayer
 from gpflux.models import DeepGP, OrthDeepGP
